Adding in HEIC conversion to jpg for the 3 different endpoints. licenseUpload, supportingDocumentUpload, saveChangeRequestDocs. Fixed some errors that were flagged by intelliJ. singlequotes and unused values.
