+++ backend/src/components/licenseUpload.js
@@ -2,6 +2,7 @@
 const {postApplicationDocument, getApplicationDocument, deleteDocument, patchOperationWithObjectId,updateChangeRequestNewFacility} = require('./utils');
 const HttpStatus = require('http-status-codes');
 const log = require('./logger');
+const {getFileExtension, convertHeicDocumentToJpg} = require('../util/common');
 
 
 async function saveLicenses(req, res) {
@@ -9,12 +10,18 @@ async function saveLicenses(req, res) {
 
     let licenses = req.body.fileList;
     for (let license of licenses) {
+
+      if (getFileExtension(license.filename) === 'heic' ) {
+        log.verbose(`saveLicenses :: heic detected for file name ${license.filename} starting conversion`);
+        license = await convertHeicDocumentToJpg(license);
+      }
+
       let response = await postApplicationDocument(license);
 
       //bind the license to the change Request Action object so the Ministry can easily see all files related to the change Action.
       if (license.changeRequestNewFacilityId){
-        let resp = await updateChangeRequestNewFacility(license.changeRequestNewFacilityId,
-          {"ccof_Attachments@odata.bind": `/ccof_application_facility_documents(${response.applicationFacilityDocumentId})`}
+        await updateChangeRequestNewFacility(license.changeRequestNewFacilityId,
+          {'ccof_Attachments@odata.bind': `/ccof_application_facility_documents(${response.applicationFacilityDocumentId})`}
         );
       }
     }
